// frontend/src/app/core/services/api/model/loginRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Credentials used for login
 */
export interface LoginRequest { 
    username: string;
    password: string;
}