// frontend/src/app/core/services/api/model/tokenResponse.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Access/refresh token response
 */
export interface TokenResponse { 
    accessToken: string;
    refreshToken: string;
    tokenType: string;
}